+++ chevah/compat/tests/normal/test_filesystem.py
@@ -1715,6 +1715,25 @@ class TestLocalFilesystem(DefaultFilesystemTestCase):
             if a_file:
                 a_file.close()
 
+    def test_openFileForReading_no_delete_lock(self):
+        """
+        A file opened only for reading will not be locked for delete
+        operations.
+        """
+        _, segments = mk.fs.makePathInTemp()
+        mk.fs.writeFileContent(segments=segments, content='something-\N{sun}')
+        a_file = None
+        try:
+            a_file = self.filesystem.openFileForReading(segments)
+
+            self.filesystem.deleteFile(segments)
+
+            content = a_file.read(100)
+            self.assertEqual(b'something-\xe2\x98\x89', content)
+        finally:
+            if a_file:
+                a_file.close()
+
     def test_openFileForWriting_ascii(self):
         """
         Check opening a file for writing in plain/ascii/str mode.